% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRate.R
\name{addRate}
\alias{addRate}
\title{Title addRate}
\usage{
addRate(dat, lab_rate, lab_cycles)
}
\arguments{
\item{dat}{matrix}

\item{lab_rate}{rate that corresponds to the cycles of the analysis performed}

\item{lab_cycles}{number of the cycles of the analysis performed}
}
\value{
the input matrix with one added column with the values of the rate of the cycles performed
}
\description{
add to the thermogram the value of rate(s) for each cycle(s) as provided by the user
}
