% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutSelect.R
\name{cutSelect}
\alias{cutSelect}
\title{Title cutSelect}
\usage{
cutSelect(x, i.start, i.end)
}
\arguments{
\item{x}{x to be cut}

\item{i.start}{index value of the starting point for the cut to be performed}

\item{i.end}{index value of the ending point for the cut to be performed}
}
\description{
cut a region of a spectra and substitutes it with a sequence with initial value i.start and end valye i.end
}
\examples{
npoints=1000
x=seq(1,npoints)
y=(dnorm(x, mean=npoints/2, sd=npoints/10))
ycut=cutValue(y,10,40,0.003,0.001)
plot(y)
lines(ycut,col="red")

}
