% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadx.R
\name{dadx}
\alias{dadx}
\title{Title dadx}
\usage{
dadx(x, a, d.step = 2)
}
\arguments{
\item{x}{denominator variable for calculating da}

\item{a}{numerator variable for calculating dt}

\item{d.step}{step of differentiation}
}
\value{
ratio of two differential of the tw0 input variables
}
\description{
calculates the ratio of two differential according to the value of d.step
}
\examples{
npoints=100
seed=42
x1=round(runif(npoints,0,1), 2)
seed=1234
x2=round(runif(npoints,0,1), 2)
xdiff <- dadx(x1,x2)
}
