% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matzero.R
\name{matzero}
\alias{matzero}
\title{Title matzero}
\usage{
matzero(mat, spks = 1, x = mat$time.seconds.zero, colname = "v.check",
  myby = "id_cycle")
}
\arguments{
\item{mat}{matrix of spectra}

\item{spks}{number of the peak selected as the starting point}

\item{x}{variable to be reset according to the position of the selected peak}

\item{colname}{name of the selected column}

\item{myby}{varialbe selected for subsetting the matrix}
}
\description{
zeroes time (in seconds) according to peak given by the user
}
