% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dev.r
\name{plot_avrami}
\alias{plot_avrami}
\title{Title plot.avrami}
\usage{
plot_avrami(out, skip = 2)
}
\arguments{
\item{out}{output from avrami function}

\item{skip}{plot symbols every nth points}
}
\description{
template for plotting results from avrami function
}
