% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dev.r
\name{plot_fri}
\alias{plot_fri}
\title{Title plot.fri}
\usage{
plot_fri(out)
}
\arguments{
\item{out}{from friedman function}
}
\description{
template for plotting results from friedman function
}
