% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dev.r
\name{plot_lavrami}
\alias{plot_lavrami}
\title{Title plot lavrami}
\usage{
plot_lavrami(out, skip = 2)
}
\arguments{
\item{out}{output from lavrami function}

\item{skip}{plot symbols every nth points}
}
\description{
template for plotting results from avrami function
}
