% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dev.r
\name{plot_ozawa}
\alias{plot_ozawa}
\title{Title plot.ozawa}
\usage{
plot_ozawa(out)
}
\arguments{
\item{out}{from ozawa function}
}
\description{
template for plotting results from avrami function
}
