% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbac.r
\name{sbAC}
\alias{sbAC}
\title{Title sbAC}
\usage{
sbAC(time.start = 0, T0 = 0, T.end = 500, qqq = 50, A = 10^(6.3),
  Ea = 80000, m = 1, n = 2, npoints = 10000,
  prec = 10^(-4.30095790876), ...)
}
\arguments{
\item{time.start}{Starting time for the simulations}

\item{T0}{Temperature start}

\item{T.end}{End temperature}

\item{qqq}{Heating rate}

\item{A}{Parameters in the equation}

\item{Ea}{Parameters in the equation}

\item{m}{Parameter in the equation}

\item{n}{Parameter in the equation}

\item{npoints}{Number of points}

\item{prec}{Starting value for the equation "prec"}

\item{...}{Parameters to pass to ode function for choosing solver method}
}
\value{
startgin temperature "T","fi",degree of crystallization "alfa",differential alfa in T "dadT",time in seconds "time.s",differential equation solution "sol"
}
\description{
Performs Šesták-Berggren (AC) simulations
}
\examples{
 \donttest{

res <- sbAC(npoints=5000,prec=10^(-4.30095790876))

}
}
\references{
J. Šesták. Thermophysical Properties of Solids, Their Measurements and Theoretical Analysis. Elsevier: Amsterdam, 1984.
}
