% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDegree.r
\name{select_degree}
\alias{select_degree}
\title{Title}
\usage{
select_degree(mat, degree = seq(0.01, 0.99, by = 0.01))
}
\arguments{
\item{mat}{matrix of the all the thermograms checked using the functiom mat.check}

\item{degree}{selected degrees of  cristallinity for performing the analysis}
}
\value{
"DT" built with the values of mat according to the specified degrees
}
\description{
Title
}
