% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simG.R
\name{simG}
\alias{simG}
\title{Title simG}
\usage{
simG(vlen, i.start, gheight, shift = 0, wd = 30)
}
\arguments{
\item{vlen}{desired length of the spectra}

\item{i.start}{starting value for the peak}

\item{gheight}{height value}

\item{shift}{shift from 0}

\item{wd}{width of the gaussian curve}
}
\description{
create a simulated spectra with gaussian shape
}
\examples{
y=(simG(500,35,1,0,w=20))
plot(y)

}
