% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTables.r
\name{summaryTableA}
\alias{summaryTableA}
\title{table.avrami}
\usage{
summaryTableA(mat.mod)
}
\arguments{
\item{mat.mod}{output matrix from avrami function}
}
\value{
table with the summary of result of applying the avrami function on the selected thermograms
}
\description{
examples of functions for presenting the results obtained with different methods
Title summaryTable A
}
