% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTables.r
\name{summaryTableFri}
\alias{summaryTableFri}
\title{Title summaryTableF}
\usage{
summaryTableFri(mat.mod)
}
\arguments{
\item{mat.mod}{output matrix from friedman function}
}
\value{
table with the summary of result of applying the friedman function on the selected thermograms
}
\description{
Title summaryTableF
}
