% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTables.r
\name{summaryTableKiss}
\alias{summaryTableKiss}
\title{Title summaryTableK}
\usage{
summaryTableKiss(mat.mod)
}
\arguments{
\item{mat.mod}{output matrix from Starink function}
}
\value{
table with the summary of result of applying the starink function on the selected thermograms
}
\description{
Title summaryTableK
}
