% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTables.r
\name{summaryTableMo}
\alias{summaryTableMo}
\title{Title summaryTable Mo}
\usage{
summaryTableMo(mat.mod)
}
\arguments{
\item{mat.mod}{output matrix from Mo function}
}
\value{
table with the summary of result of applying the ozawwa function on the selected thermograms
}
\description{
Title summaryTable Mo
}
