% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTables.r
\name{summaryTableOz}
\alias{summaryTableOz}
\title{Title summaryTableOz}
\usage{
summaryTableOz(mat.mod)
}
\arguments{
\item{mat.mod}{output matrix from ozawa function}
}
\value{
table with the summary of result of applying the ozawwa function on the selected thermograms
}
\description{
Title summaryTableOz
}
