% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbaseline.R
\name{t_baseline}
\alias{t_baseline}
\title{Title t.baseline}
\usage{
t_baseline(y)
}
\arguments{
\item{y}{baseline correction on y}
}
\value{
y.baseline returns the corrected y
}
\description{
a wrapper for the baseline.rdfbaseline function in the package baseline in order to have the output in the same format as the input
}
\examples{
\donttest{
y.baseline <- t_baseline(y)
}
}
