% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMat.R
\name{testMat}
\alias{testMat}
\title{Title testMat}
\usage{
testMat(a, l.lim = 1, r.lim = NULL, toselect = c(0, 1, 2, 0, 0, 0, 3, 4))
}
\arguments{
\item{a}{list of data tables of the checked thermograms using checkmat , obtained at different rates to change lines}

\item{l.lim}{left lim of running integral}

\item{r.lim}{rigth lim of running integral}

\item{toselect}{vector}
}
\value{
data table ready to be used by all the methods for kinetic analysis included in the package
}
\description{
Title testMat
}
\examples{
require(data.table)
npoints=1000
x=seq(1,npoints)
y=(dnorm(x, mean=npoints/2, sd=npoints/10))
x=seq(1,1000)
x2=seq(200,500,length.out=1000)
dat=data.frame(x,x2,y)
colnames(dat) <- c("time.seconds", "temperature.s","heat.flow")
dat=data.table(dat)
dat2=dat
dat$rates=20
dat2$rates=50
toTest=list(dat,dat2)
tested=testMat(toTest)
}
